feat(onboarding): move onboarding slides
- move all slides from @/components/onboarding to @/components/onboarding/slides
