+++ screenpipe-app-tauri/components/onboarding/slides/api-setup.tsx
@@ -6,9 +6,9 @@ import { ArrowUpRight } from "lucide-react";
 import { DialogHeader, DialogTitle } from "@/components/ui/dialog";
 import { open } from "@tauri-apps/plugin-shell";
 
-import OnboardingNavigation from "@/components/onboarding/navigation";
+import OnboardingNavigation from "@/components/onboarding/slides/navigation";
 
-import AISection from "../settings/ai-section";
+import AISection from "../../settings/ai-section";
 
 interface OnboardingAPISetupProps {
   className?: string;
