+++ screenpipe-app-tauri/components/onboarding/slides/login.tsx
