+++ screenpipe-app-tauri/components/onboarding/slides/navigation.tsx
