+++ screenpipe-app-tauri/components/onboarding/slides/pipe-store.tsx
