+++ screenpipe-app-tauri/components/onboarding/slides/single-codeblock.tsx
